+++ examples/terraform/main.tf
@@ -1,12 +1,12 @@
 # =======================================================================
 # Terraform Vault Provider example configuration
 #
-# NB: This configuration performs all of the same post-unseal steps that
-#     `blazing_sword` performs.
-#
-# - Establish a file based audit log device resource
-# -
+# NB: This Terraform configuration performs all post-unseal setup
+#     See README.md for more details
 # ========================================================================
+
+# We presume Vault at https://localhost:8200
+# and the presence of ~/.vault-token here.
 
 provider "vault" {
 
@@ -29,7 +29,7 @@ resource "vault_audit" "vaultron_audit_device" {
 
   options = {
     file_path = "/vault/logs/audit.log"
-    description = "Socket device by Vaultron and Terraform Vault provider"
+    description = "Vaultron example file audit device"
   }
 }
 
@@ -40,25 +40,25 @@ resource "vault_audit" "vaultron_audit_device" {
 resource "vault_auth_backend" "vaultron_approle" {
   type = "approle"
   path = "vaultron-approle"
-  description = "AppRole auth method by Vaultron and Terraform Vault provider"
+  description = "Vaultron example AppRole auth method"
 }
 
 resource "vault_auth_backend" "vaultron_cert" {
   type = "cert"
   path = "vaultron-cert"
-  description = "X.509 certificate auth method by Vaultron and Terraform Vault provider"
+  description = "Vaultron example X.509 certificate auth method"
 }
 
 resource "vault_auth_backend" "vaultron_userpass" {
   type = "userpass"
   path = "vaultron-userpass"
-  description = "Username and password auth method by Vaultron and Terraform Vault provider"
+  description = "Vaultron example Username and password auth method"
 }
 
 resource "vault_auth_backend" "vaultron_ldap" {
   type = "ldap"
   path = "vaultron-ldap"
-  description = "LDAP auth method by Vaultron and Terraform Vault provider"
+  description = "Vaultron example LDAP auth method"
 }
 
 
@@ -69,47 +69,119 @@ resource "vault_auth_backend" "vaultron_ldap" {
 resource "vault_mount" "vaultron_kv" {
   path        = "vaultron-kv"
   type        = "kv"
-  description = "KV version 1 Secrets Engine by Vaultron and Terraform Vault provider"
+  description = "Vaultron example KV version 1 secrets engine"
 }
 
 resource "vault_mount" "vaultron_aws" {
   path        = "vaultron-aws"
   type        = "aws"
-  description = "AWS Secrets Engine by Vaultron and Terraform Vault provider"
+  description = "Vaultron example AWS secrets engine"
 }
 
 resource "vault_mount" "vaultron_consul" {
   path        = "vaultron-consul"
   type        = "consul"
-  description = "Consul Secrets Engine by Vaultron and Terraform Vault provider"
+  description = "Vaultron example Consul secrets engine"
 }
 
 resource "vault_mount" "vaultron_pki_root" {
   path        = "vaultron-root-pki"
   type        = "pki"
-  description = "PKI Secrets Engine (root CA) by Vaultron and Terraform Vault provider"
+  description = "Vaultron example PKI secrets engine (for root CA)"
 }
 
 resource "vault_mount" "vaultron_pki_int" {
   path        = "vaultron-root-int"
   type        = "pki"
-  description = "PKI Secrets Engine (int CA) by Vaultron and Terraform Vault provider"
+  description = "Vaultron example PKI secrets engine (for int CA)"
 }
 
 resource "vault_mount" "vaultron_transit" {
   path        = "vaultron-transit"
   type        = "transit"
-  description = "Transit Secrets Engine by Vaultron and Terraform Vault provider"
+  description = "Vaultron example Transit secrets engine"
 }
 
 resource "vault_mount" "vaultron_ssh_host_signer" {
   path        = "vaultron-ssh-host-signer"
   type        = "ssh"
-  description = "SSH Secrets Engine (host) by Vaultron and Terraform Vault provider"
+  description = "Vaultron example SSH Secrets Engine (host)"
 }
 
 resource "vault_mount" "vaultron_ssh_client_signer" {
   path        = "vaultron-ssh-client-signer"
   type        = "ssh"
-  description = "SSH Secrets Engine (client) by Vaultron and Terraform Vault provider"
+  description = "Vaultron example SSH Secrets Engine (client)"
+}
+
+
+# -----------------------------------------------------------------------
+# Policy Resources
+# -----------------------------------------------------------------------
+
+resource "vault_policy" "vaultron_wildcard" {
+  name = "wildcard"
+  policy = <<EOT
+// Vaultron example policy: "vaultron-wildcard"
+path "*" {
+  capabilities = ["create", "read", "update", "delete", "list", "sudo"]
+}
+EOT
+}
+
+resource "vault_policy" "vaultron_example_root_ns" {
+  name = "vaultron-example-root-ns"
+  policy = <<EOT
+// Vaultron example policy: "example root namespace"
+# Manage namespaces
+
+path "sys/namespaces/*" {
+  capabilities = ["create", "read", "update", "delete", "list", "sudo"]
+}
+
+# Manage policies via API
+path "sys/policies/*" {
+  capabilities = ["create", "read", "update", "delete", "list", "sudo"]
+}
+
+# Manage policies via CLI
+path "sys/policy/*" {
+  capabilities = ["create", "read", "update", "delete", "list", "sudo"]
+}
+
+# List policies via CLI
+path "sys/policy" {
+  capabilities = ["read", "update", "list"]
+}
+
+# Enable and manage secrets engines
+path "sys/mounts/*" {
+  capabilities = ["create", "read", "update", "delete", "list"]
+}
+
+# List available secret engines
+path "sys/mounts" {
+  capabilities = [ "read" ]
+}
+
+# Create and manage entities and groups
+path "identity/*" {
+  capabilities = ["create", "read", "update", "delete", "list"]
+}
+
+# Manage tokens
+path "auth/token/*" {
+  capabilities = ["create", "read", "update", "delete", "list", "sudo"]
+}
+EOT
+}
+
+resource "vault_policy" "vaultron_example_ns" {
+  name = "vaultron-example-ns"
+  policy = <<EOT
+// Vaultron example policy: "example vaultron namespace"
+path "sys/namespaces/vaultron/*" {
+  capabilities = ["create", "read", "update", "delete", "list", "sudo"]
+}
+EOT
 }